Fixed reading of large files within ISO (file size >= 0x80000000).

+++ src/jpcsp/filesystems/umdiso/iso9660/Iso9660Handler.java
@@ -45,7 +45,7 @@ public class Iso9660Handler extends Iso9660Directory {
         Iso9660File rootDirEntry = new Iso9660File(b,b.length);
 
         int rootLBA = rootDirEntry.getLBA();
-        int rootSize = rootDirEntry.getSize();
+        long rootSize = rootDirEntry.getSize();
 
         internalDir = new Iso9660Directory(r, rootLBA, rootSize);
     }
